refactor(gameplay): unify DiceRoller and D6Roller types into shared diceTypes module
- Created src/utils/gameplay/diceTypes.ts as single source of truth
- Moved D6Roller, DiceRoller types and dice rolling functions from hitLocation.ts
- Removed duplicate DiceRoller definitions from gameSession.ts and specialWeaponMechanics.ts
- Maintained backward compatibility via re-exports from original locations
- All 19,145 tests pass, zero TypeScript errors
