+++ src/utils/gameplay/hitLocation.ts
@@ -11,6 +11,8 @@ import {
   IDiceRoll,
   IHitLocationResult,
 } from '@/types/gameplay';
+export { type D6Roller, defaultD6Roller, rollD6, roll2d6 } from './diceTypes';
+import { type D6Roller, defaultD6Roller, rollD6, roll2d6 } from './diceTypes';
 
 // =============================================================================
 // Hit Location Tables
@@ -113,35 +115,8 @@ export function getHitLocationTable(
 }
 
 // =============================================================================
-// Dice Rolling
+// Dice Rolling (re-exported from diceTypes.ts — single source of truth)
 // =============================================================================
-
-export type D6Roller = () => number;
-
-const defaultD6Roller: D6Roller = () => Math.floor(Math.random() * 6) + 1;
-
-/**
- * Roll a single d6. Accepts an optional injectable roller for deterministic testing.
- */
-export function rollD6(roller: D6Roller = defaultD6Roller): number {
-  return roller();
-}
-
-/**
- * Roll 2d6 and return detailed result. Accepts an optional injectable roller.
- */
-export function roll2d6(roller: D6Roller = defaultD6Roller): IDiceRoll {
-  const die1 = rollD6(roller);
-  const die2 = rollD6(roller);
-  const total = die1 + die2;
-
-  return {
-    dice: [die1, die2],
-    total,
-    isSnakeEyes: total === 2,
-    isBoxcars: total === 12,
-  };
-}
 
 /**
  * Create a dice roll result from a predetermined value (for testing/replays).
